reduce some casts

+++ programs/clearing_house/src/controller/pnl/delisting.rs
@@ -2370,7 +2370,7 @@ pub mod delisting_test {
 
             liquidate_perp(
                 0,
-                shorter.perp_positions[0].base_asset_amount.unsigned_abs() as u128,
+                shorter.perp_positions[0].base_asset_amount.unsigned_abs(),
                 &mut shorter,
                 &maker_key,
                 &mut shorter_user_stats,
