+++ programs/clearing_house/src/validation/order.rs
@@ -40,7 +40,7 @@ pub fn validate_order(
     Ok(())
 }
 
-fn validate_market_order(order: &Order, step_size: u128) -> ClearingHouseResult {
+fn validate_market_order(order: &Order, step_size: u64) -> ClearingHouseResult {
     validate_base_asset_amount(order, step_size)?;
 
     match order.direction {
@@ -170,7 +170,7 @@ fn validate_post_only_order(
 
 fn validate_trigger_limit_order(
     order: &Order,
-    step_size: u128,
+    step_size: u64,
     minimum_order_value: u128,
 ) -> ClearingHouseResult {
     validate_base_asset_amount(order, step_size)?;
@@ -226,7 +226,7 @@ fn validate_trigger_limit_order(
 
 fn validate_trigger_market_order(
     order: &Order,
-    step_size: u128,
+    step_size: u64,
     minimum_order_value: u128,
 ) -> ClearingHouseResult {
     validate_base_asset_amount(order, step_size)?;
@@ -266,14 +266,14 @@ fn validate_trigger_market_order(
     Ok(())
 }
 
-fn validate_base_asset_amount(order: &Order, step_size: u128) -> ClearingHouseResult {
+fn validate_base_asset_amount(order: &Order, step_size: u64) -> ClearingHouseResult {
     if order.base_asset_amount == 0 {
         msg!("Order base_asset_amount cant be 0");
         return Err(ErrorCode::InvalidOrder);
     }
 
     validate!(
-        is_multiple_of_step_size(order.base_asset_amount as u128, step_size)?,
+        is_multiple_of_step_size(order.base_asset_amount, step_size)?,
         ErrorCode::InvalidOrder,
         "Order base asset amount ({}) not a multiple of the step size ({})",
         order.base_asset_amount,
@@ -287,7 +287,7 @@ pub fn validate_spot_order(
     order: &Order,
     valid_oracle_price: Option<i128>,
     slot: u64,
-    step_size: u128,
+    step_size: u64,
     margin_ratio_initial: u128,
     margin_ratio_maintenance: u128,
     minimum_order_value: u128,
@@ -320,7 +320,7 @@ fn validate_spot_limit_order(
     order: &Order,
     valid_oracle_price: Option<i128>,
     slot: u64,
-    step_size: u128,
+    step_size: u64,
     margin_ratio_initial: u128,
     margin_ratio_maintenance: u128,
     minimum_order_value: u128,
